re #99151.
